+++ server/init.sql
@@ -54,18 +54,38 @@ CREATE UNIQUE INDEX service_id_str_unique
  ON public.service
  ( id_str );
 
-CREATE SEQUENCE public.needs_update_id_seq;
+CREATE SEQUENCE public.user_var_needs_update_id_seq;
 
-CREATE TABLE public.needs_update (
-                id INTEGER NOT NULL DEFAULT nextval('public.needs_update_id_seq'),
-                service_trigger_id INTEGER NOT NULL,
-                service_updated_id INTEGER NOT NULL,
-                last_ip VARCHAR(15) NOT NULL,
-                CONSTRAINT needs_update_pk PRIMARY KEY (id)
+CREATE TABLE public.user_var_needs_update (
+                id INTEGER NOT NULL DEFAULT nextval('public.user_var_needs_update_id_seq'),
+                service_id INTEGER NOT NULL,
+                user_variable_id INTEGER NOT NULL,
+                last_value VARCHAR,
+                CONSTRAINT user_var_needs_update_pk PRIMARY KEY (id)
 );
 
 
-ALTER SEQUENCE public.needs_update_id_seq OWNED BY public.needs_update.id;
+ALTER SEQUENCE public.user_var_needs_update_id_seq OWNED BY public.user_var_needs_update.id;
+
+CREATE SEQUENCE public.ip_needs_update_id_seq;
+
+CREATE TABLE public.ip_needs_update (
+                id INTEGER NOT NULL DEFAULT nextval('public.ip_needs_update_id_seq'),
+                service_trigger_id INTEGER NOT NULL,
+                service_updated_id INTEGER NOT NULL,
+                last_ip VARCHAR(15) NOT NULL,
+                CONSTRAINT ip_needs_update_pk PRIMARY KEY (id)
+);
+
+
+ALTER SEQUENCE public.ip_needs_update_id_seq OWNED BY public.ip_needs_update.id;
+
+ALTER TABLE public.user_var_needs_update ADD CONSTRAINT user_variable_user_var_needs_update_fk
+FOREIGN KEY (user_variable_id)
+REFERENCES public.user_variable (id)
+ON DELETE NO ACTION
+ON UPDATE NO ACTION
+NOT DEFERRABLE;
 
 ALTER TABLE public.service ADD CONSTRAINT server_service_fk
 FOREIGN KEY (server_id)
@@ -74,15 +94,22 @@ ON DELETE NO ACTION
 ON UPDATE NO ACTION
 NOT DEFERRABLE;
 
-ALTER TABLE public.needs_update ADD CONSTRAINT service_needs_update_fk
+ALTER TABLE public.ip_needs_update ADD CONSTRAINT service_needs_update_fk
 FOREIGN KEY (service_trigger_id)
 REFERENCES public.service (id)
 ON DELETE NO ACTION
 ON UPDATE NO ACTION
 NOT DEFERRABLE;
 
-ALTER TABLE public.needs_update ADD CONSTRAINT service_needs_update_fk1
+ALTER TABLE public.ip_needs_update ADD CONSTRAINT service_needs_update_fk1
 FOREIGN KEY (service_updated_id)
+REFERENCES public.service (id)
+ON DELETE NO ACTION
+ON UPDATE NO ACTION
+NOT DEFERRABLE;
+
+ALTER TABLE public.user_var_needs_update ADD CONSTRAINT service_user_var_needs_update_fk
+FOREIGN KEY (service_id)
 REFERENCES public.service (id)
 ON DELETE NO ACTION
 ON UPDATE NO ACTION