+++ server/src/config/emergency.py
@@ -4,7 +4,7 @@ import os
 from config.load import get_config
 from logger import logger
 from config.emergency_procedure import EmergencyProcedure
-
+from config.parser import parse_json_file
 
 class EmergencyProceduresConfig:
     def __init__(self) -> None:
@@ -17,7 +17,7 @@ class EmergencyProceduresConfig:
         if emergency_config is None:
             normal_config = get_config()
             emergency_config = os.path.join(
-                os.path.dirname(normal_config), "emergency_proc.json"
+                os.path.dirname(normal_config), "emergency_proc.jsonc"
             )
         self.config_path = emergency_config
 
@@ -25,8 +25,7 @@ class EmergencyProceduresConfig:
             logger.warning(f"Config not found ({emergency_config})")
             self.config_raw = {}
         else:
-            with open(emergency_config, "r", encoding="utf-8") as f:
-                self.config_raw = json.load(f)
+            self.config_raw = parse_json_file(emergency_config)
 
     def apply_config(self):
         for event_name, listener in self.event_listeners.items():