+++ server/src/plugins/commands/library.py
@@ -10,6 +10,7 @@ from plugins.commands.service_unassign import ServiceUnassignCommand
 from plugins.commands.services_list import ServiceListCommand
 from plugins.commands.services_sync import ServicesSync
 from plugins.commands.emergency_procedure import EmergencyProcedureCommand
+from plugins.commands.set_variable import SetUserVarCommand
 
 COMMANDS: list[type[CommandBase]] = [
     ConfigSyncCommand,
@@ -22,5 +23,6 @@ COMMANDS: list[type[CommandBase]] = [
     RawCommand,
     ConfigRuntimeCommand,
     RegenConfigCommand,
-    EmergencyProcedureCommand
+    EmergencyProcedureCommand,
+    SetUserVarCommand,
 ]